app/requirements.txt
torch
torchvision
torchaudio
numpy
pillow
opencv-python-headless
matplotlib
scikit-learn
requests
streamlit
